Revert "[Refactor] Extract RerouteNode as a separate file (#2668)"
This reverts commit 85d04f68145633a3c0ae715595f10cdd7c369b20.
