+++ src/extensions/core/groupNode.ts
@@ -1,12 +1,8 @@
 // @ts-strict-ignore
-import { LGraphCanvas, LiteGraph } from '@comfyorg/litegraph'
+import { type LGraph, LGraphCanvas, LiteGraph } from '@comfyorg/litegraph'
 import { LGraphNode, type NodeId } from '@comfyorg/litegraph/dist/LGraphNode'
 
-import { ManageGroupDialog } from '@/extensions/core/groupNodeManage'
 import { t } from '@/i18n'
-import { mergeIfValid } from '@/nodes/PrimitiveNode'
-import { api } from '@/scripts/api'
-import { app } from '@/scripts/app'
 import { useDialogService } from '@/services/dialogService'
 import { useNodeDefStore } from '@/stores/nodeDefStore'
 import { useToastStore } from '@/stores/toastStore'
@@ -14,6 +10,11 @@ import type { ComfyNodeDef } from '@/types/apiTypes'
 import { ComfyExtension } from '@/types/comfy'
 import { ComfyLink, ComfyNode, ComfyWorkflowJSON } from '@/types/comfyWorkflow'
 import { deserialiseAndCreate, serialise } from '@/utils/vintageClipboard'
+
+import { api } from '../../scripts/api'
+import { app } from '../../scripts/app'
+import { ManageGroupDialog } from './groupNodeManage'
+import { mergeIfValid } from './widgetInputs'
 
 type GroupNodeWorkflowData = {
   external: ComfyLink[]
